SISRP-45434 - Post-3/3 React Cleanup - consolidate Spinner components
