+++ src/react/components/Spinner.js
@@ -1,16 +1,31 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import '../stylesheets/widgets.scss';
+import '../stylesheets/spinners.scss';
+
 const propTypes = {
-  children: PropTypes.node
+  isLoadingMessage: PropTypes.string
 };
 
-const Spinner = (props) => (
-  <div className="cc-spinner" aria-live="polite" aria-busy={true}>
-    {props.children && <p className="cc-spinner-message">{props.children}</p>}
-  </div>
-);
+const renderMessage = (isLoadingMessage) => {
+  if (isLoadingMessage) {
+    return (
+      <p className="cc-react-spinner-message">{isLoadingMessage}</p>
+    );
+  } else {
+    return null;
+  }
+};
 
+const Spinner = (props) => {
+  return (
+    <div className='cc-react-widget--padding'>
+      <div aria-live="polite" className="cc-react-spinner"></div>
+      {renderMessage(props.isLoadingMessage)}
+    </div>
+  );
+};
 Spinner.propTypes = propTypes;
 
 export default Spinner;
